fix(logging): move task output logging and improve output quality
