+++ lib/get/get-full-source-space.js
@@ -1,35 +1,12 @@
 import Promise from 'bluebird'
-import figures from 'figures'
 import Listr from 'listr'
 import verboseRenderer from 'listr-verbose-renderer'
 
-import { logEmitter, formatErrorOneLine } from '../utils/logging'
+import { logEmitter, logToTaskOutput } from '../utils/logging'
 import sortEntries from '../utils/sort-entries'
 
 const MAX_ALLOWED_LIMIT = 1000
 let pageLimit = MAX_ALLOWED_LIMIT
-
-function logToTaskOutput (task) {
-  function logInfo (info) {
-    task.output = `${figures.tick} ${info}`
-  }
-  function logWarning (warning) {
-    task.output = `${figures.warning} ${warning}`
-  }
-  function logError (error) {
-    const formattedError = formatErrorOneLine(error)
-    task.output = `${figures.cross} ${formattedError}`
-  }
-  logEmitter.on('info', logInfo)
-  logEmitter.on('warning', logWarning)
-  logEmitter.on('error', logError)
-
-  return () => {
-    logEmitter.removeListener('info', logInfo)
-    logEmitter.removeListener('warning', logWarning)
-    logEmitter.removeListener('error', logError)
-  }
-}
 
 /**
  * Gets all the content from a space via the management API. This includes
