+++ lib/utils/get-entity-name.js
@@ -2,17 +2,28 @@ import { get } from 'lodash'
 
 export default function getEntityName (entity) {
   const name = get(entity, 'name')
-  const titleField = get(entity, 'fields.title')
-  const id = get(entity, 'sys.id')
   if (name) {
-    return name
+    return attachId(name, entity)
   }
+
+  const titleField = get(entity, 'fields.title')
   if (titleField) {
     const locales = Object.keys(titleField)
-    return titleField[locales[0]]
+    return attachId(titleField[locales[0]], entity)
   }
+
+  const id = get(entity, 'sys.id')
   if (id) {
     return id
   }
+
   return 'unknown'
+}
+
+function attachId (val, entity) {
+  const id = get(entity, 'sys.id')
+  if (id) {
+    return `${val} (${id})`
+  }
+  return val
 }
