+++ lib/utils/logging.js
@@ -2,6 +2,7 @@ import EventEmitter from 'events'
 
 import bfj from 'bfj-node4'
 import { isObject } from 'lodash'
+import figures from 'figures'
 import moment from 'moment'
 
 export const logEmitter = new EventEmitter()
@@ -40,6 +41,9 @@ export function formatErrorOneLine (error) {
     }
     return `${error.error.name}: ${errorOutput.join(' - ')}`
   } catch (err) {
+    if (!error.hasOwnProperty(('error'))) {
+      error = {error}
+    }
     return error.error.toString().replace(/[\n\r]/, ' ')
   }
 }
@@ -108,4 +112,26 @@ export function setupErrorLogging (errorLog) {
 
   logEmitter.addListener('warning', (error) => errorLogger('warning', error))
   logEmitter.addListener('error', (error) => errorLogger('error', error))
+}
+
+export function logToTaskOutput (task) {
+  function logInfo (info) {
+    task.output = `${figures.tick} ${info}`
+  }
+  function logWarning (warning) {
+    task.output = `${figures.warning} ${warning}`
+  }
+  function logError (error) {
+    const formattedError = formatErrorOneLine(error)
+    task.output = `${figures.cross} ${formattedError}`
+  }
+  logEmitter.on('info', logInfo)
+  logEmitter.on('warning', logWarning)
+  logEmitter.on('error', logError)
+
+  return () => {
+    logEmitter.removeListener('info', logInfo)
+    logEmitter.removeListener('warning', logWarning)
+    logEmitter.removeListener('error', logError)
+  }
 }
